+++ CHANGES.md
@@ -12,6 +12,7 @@ Irccd:
   standard paths for both irccd and plugins (#611),
 - If Mercurial is found, the version is bundled in `irccd --version`,
 - Irccd no longer supports uid, gid, pid and daemon features (#846).
+- Sections `[identity]` and `[server]` have been merged (#905).
 
 Irccd test: